Fan control scripts for ZyXEL NBG7815
The very first commit here is the extracted ipk package for the fancontrol provided
by ZyXEL for this device. It is part of the GPL archive provided by ZyXEL for this
device.

The following commits are a rework of the scripts to make it working on OpenWrt.

Attached to this initial commit is the original ipk package provided by ZyXEL.

Installation:

`opkg install fanctld_v*_aarch64_cortex-a53.ipk`

The configuration file /etc/config/fanctld will not be overwritten. If you want to
do you have to use:

`opkg install --force-maintainer fanctld_v*_aarch64_cortex-a53.ipk`

To use this package you have to make your own build of OpenWrt for this device.
You have further to modify the config:

- add the sensors to the DTS file
- add the fan to the DTS file
- add the necessary drivers (kmod-hwmon-tmp103, kmod-hwmon-gpiofan) to config

If you do your own builds anyway you don't need the ipk package. It is enough
to integrate the content of the `data` directory into the `files` directory.